backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlmodel
passlib[bcrypt]
pyjwt
httpx
pydantic
